Using generic parsing for Table Entities with json (task #5394)
Mapping of the data is simplified to parse the incoming data.
We assume that the user adds a Trait to Model Table classes,
that simplify the convertion to Generic ObjectType Interface.

It'll map the incoming data into the object. Later it'll be
converted to ORM\Entity and stored in the system.